+++ src/systems/si/dimensions.rs
@@ -120,7 +120,9 @@ pub type SpecificHeatCapacity = Dimension<P2, Z0, N2, Z0, N1, Z0, Z0>;
 /// Thermal conductivity.
 pub type ThermalConductivity = Dimension<P1, P1, N3, Z0, N1, Z0, Z0>;
 /// Thermal diffusivity.
-pub type ThermalDiffusivity = Dimension<P2, Z0, N1, Z0, Z0, Z0, Z0>;
+pub type ThermalDiffusivity = AreaPerTime;
+/// Area per time, including planar flow rate per unit width.
+pub type AreaPerTime = Dimension<P2, Z0, N1, Z0, Z0, Z0, Z0>;
 /// Mass density.
 pub type MassDensity = Dimension<N3, P1, Z0, Z0, Z0, Z0, Z0>;
 /// Mass-density rate, mass per volume per time.